requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
asyncpg==0.29.0

alembic==1.14.0

python-jose[cryptography]==3.3.0
pydantic-settings==2.7.0
